requirements.txt
# Python 3.11+ recommended (tested on 3.11.14)
streamlit==1.30.0
Pillow>=10.0.0
google-generativeai>=0.4.0
numpy>=1.26.0
python-dotenv>=1.0.0

# FAISS Vector Store for Hybrid RAG (Phase 1.5)
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
pypdf>=3.17.0

# ADK (Google Agent Development Kit) - Real integration with google-adk SDK
google-adk==1.19.0
google-genai>=0.3.0
